==> 3.2_drug_cell_graph.ipynb <==
import pandas as pd
from itertools import combinations
from scipy.stats import zscore
from sklearn.preprocessing import MinMaxScaler
import networkx as nx

from networkx.algorithms import bipartite


df_small = pd.read_csv("training_data/df_small[drug-protein].csv")
df_large =pd.read_csv("training_data/df_large[drug-protein].csv")

avail_drugs_small = set(df_small['Drug1'].unique().tolist() + df_small['Drug2'].unique().tolist())
avail_drugs_large = set(df_large['Drug1'].unique().tolist() + df_large['Drug2'].unique().tolist())
avail_cells_small = df_small['Cell line'].unique().tolist()


DC = nx.Graph()

DC.add_nodes_from(avail_drugs_small)
DC.add_nodes_from(avail_cells_small)

edges1 = [(row['Drug1'], row['Cell line'])  for _, row in df_small.iterrows()]
edges2 = [(row['Drug2'], row['Cell line'])  for _, row in df_small.iterrows()]

DC.add_edges_from(edges1)
DC.add_edges_from(edges2)

df_small.dtypes

def get_all_stats(DC, avail_drugs_small, df_og):
    drug_projection = bipartite.projected_graph(DC, avail_drugs_small)
    dc_deg_centrality = nx.degree_centrality(drug_projection)
    dc_cc = nx.clustering(drug_projection)
    dc_pagerank = nx.pagerank(drug_projection)
    dc_closeness = nx.closeness_centrality(drug_projection)
    dc_betweenness = nx.betweenness_centrality(drug_projection)

    def add_node_features(df, feature_dict, feature_name):
        df["dc_" + feature_name + "_1"] = df["Drug1"].map(feature_dict)
        df["dc_" + feature_name + "_2"] = df["Drug2"].map(feature_dict)
        return df

    df_og_extra = df_og.copy()

    df_og_extra = add_node_features(
    df_og_extra, dc_deg_centrality, "degree_centrality"
    )
    df_og_extra = add_node_features(df_og_extra, dc_cc, "clustering")
    df_og_extra = add_node_features(df_og_extra, dc_pagerank, "pagerank")
    df_og_extra = add_node_features(df_og_extra, dc_closeness, "closeness")
    df_og_extra = add_node_features(df_og_extra, dc_betweenness, "betweenness")

    return df_og_extra 


df_small_extra = get_all_stats(DC, avail_drugs_small, df_small)

df_small_extra.to_csv("training_data/df_small[both].csv",index=False)

from itertools import combinations 

all_pairs = list(combinations(avail_drugs_small, 2))

existing_pairs = set(tuple(sorted([row['Drug1'], row['Drug2']])) for index, row in df_small.iterrows())
new_pairs = [pair for pair in all_pairs if pair not in existing_pairs]
new_data = []
for cell_line in avail_cells_small:
    for drug1, drug2 in new_pairs:
        new_data.append({
            'Drug1': drug1,
            'Drug2': drug2,
            'Cell line': cell_line
        })

# Create a new dataframe from the new pairs
df_new_combinations = pd.DataFrame(new_data)

df_new_combinations.to_csv("new_combinations.csv",index=False)

df_drug_protein = pd.read_csv("filtered_data/df_drug_protein_2(normalized).csv")
df_drug_cid = pd.read_csv("filtered_data/df_drug_cid.csv")


def get_avg_z(row):
    z1 = df_drug_protein[df_drug_protein["chemical"] == row[0]]["z_score"]
    z2 = df_drug_protein[df_drug_protein["chemical"] == row[1]]["z_score"]
    return (z1 + z2) / 2


def get_avg_mm(row):
    z1 = df_drug_protein[df_drug_protein["chemical"] == row["chemical_1"]][
        "min_max_score"
    ]
    z2 = df_drug_protein[df_drug_protein["chemical"] == row["chemical_2"]][
        "min_max_score"
    ]
    return (z1 + z2) / 2


def get_test_data(df_new_comb, DC, avail_drugs, sample_size=100):
    df_test = df_new_comb.sample(sample_size).reset_index()
    DC_test = DC.copy()
    edges1 = [(row["Drug1"], row["Cell line"]) for _, row in df_test.iterrows()]
    edges2 = [(row["Drug2"], row["Cell line"]) for _, row in df_test.iterrows()]
    DC_test.add_edges_from(edges1)
    DC_test.add_edges_from(edges2)
    c1 = [
        "Drug1",
        "degree_centrality_1",
        "clustering_1",
        "pagerank_1",
        "closeness_1",
        "betweenness_1",
    ]
    c2 = [
        "Drug2",
        "degree_centrality_2",
        "clustering_2",
        "pagerank_2",
        "closeness_2",
        "betweenness_2",
    ]
    df2 = df_test.merge(df_drug_cid, how="left", left_on="Drug1", right_on="Drug")[
        ["Drug1", "Drug2", "Cell line", "cIds"]
    ]
    df2 = df2.rename(columns={"cIds": "chemical_1"})
    df2 = df2.merge(df_drug_cid[["Drug", "cIds"]], left_on="Drug2", right_on="Drug")[
        ["Drug1", "Drug2", "Cell line", "chemical_1", "cIds"]
    ]
    df2 = df2.rename(columns={"cIds": "chemical_2"})
    df_drug_protein_avg = (
        df_drug_protein[["chemical", "min_max_score", "z_score", "min_max_score"]]
        .groupby("chemical")
        .mean()
        .reset_index()
    )
    df2 = df2.merge(
        df_drug_protein_avg[["chemical", "z_score", "min_max_score"]],
        how="left",
        left_on="chemical_1",
        right_on="chemical",
        suffixes=("", "_1"),
    )
    df2 = df2.merge(
        df_drug_protein_avg[["chemical", "z_score", "min_max_score"]],
        how="left",
        left_on="chemical_2",
        right_on="chemical",
        suffixes=("", "_2"),
    )
    df2["z_score_sum"] = df2[["z_score", "z_score_2"]].mean(axis=1)
    df2["min_max_score_sum"] = df2[["min_max_score", "min_max_score_2"]].mean(axis=1)
    df2.drop(
        columns=[
            "chemical",
            "z_score",
            "z_score_2",
            "chemical_2",
            "chemical_1",
            "min_max_score_2",
            "min_max_score",
        ],
        inplace=True,
    )

    df2 = df2.merge(df_small[c1], on="Drug1", how="left").drop_duplicates()
    df2 = df2.merge(df_small[c2], on='Drug2', how='left').drop_duplicates()
    df2 = df2.dropna()
    df_combined = pd.concat([df_small, test_data]).drop(
    columns=[
        "Bliss",
        "HSA",
        "Loewe",
        "ZIP",
        "chemical_1",
        "chemical_2",
        "classification",
        "combination_id"
        ]
    )

    


print(DC)
test_data = get_test_data(df_new_combinations, DC, avail_drugs_small)

set(df_small.columns) -set(test_data.columns) 

df_combined = pd.concat([df_small, test_data]).drop(
    columns=[
        "Bliss",
        "HSA",
        "Loewe",
        "ZIP",
        "chemical_1",
        "chemical_2",
        "classification",
        "combination_id"
    ]
)

df_combined

df2 = test_data.copy()



df2
